Review with Oswaldo - matlab processing script and functions fun_ProbTables and fun_MeanTables
I'd like to review the changes made here with Oswaldo- I think I understand what is going on but there is a lot of detail to pay attention to!

I appreciate the consistency for the 9 sections of this code- 3 forests and 3 models.  I made some minor updates to the file paths to make it easier for someone else to take the .matlab files and produce the excel sheets.

I also saw that there was a repeated df name for example

J_resultsModel1_Folder <- readMat("Arb_resultsModel4.mat")

The .mat file has the intended name, but the df it creates is sortof confusing.  I do see the environment being cleared at the top of each of the 9 processing sections, which is a nice safety procedure.

Lastly, I commented out each time the function was called to make pdf images. This reduced the processing time by ~2 fold, but you may know if there are things I'm missing by choosing not to run this function.